# Fix dark mode toggle performance bottlenecks