--- a/issue.md
+++ b/issue.md
@@ -1,0 +1,1 @@
+light/dark mode toggle on readings page is not performant; it is noticeably slow