--- a/src/app/components/__tests__/ThemePerformance.integration.test.tsx
+++ b/src/app/components/__tests__/ThemePerformance.integration.test.tsx
@@ -1,0 +1,302 @@
+/**
+ * Theme Performance Integration Tests
+ * @jest-environment jsdom
+ *
+ * Tests to ensure theme switching doesn't cause unnecessary re-renders,
+ * preventing regression of the N+1 subscription anti-pattern fixed in PR #56
+ */
+
+import React from 'react';
+import { render, screen } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
+import ReadingCard from '../readings/ReadingCard';
+import DarkModeToggle from '../DarkModeToggle';
+import type { ReadingListItem } from '@/types';
+
+// Mock the logger
+jest.mock('@/lib/logger', () => ({
+  logger: {
+    warn: jest.fn(),
+    error: jest.fn(),
+    info: jest.fn(),
+    debug: jest.fn(),
+  },
+}));
+
+// Mock the UI store
+const mockToggleDarkMode = jest.fn();
+const mockSetDarkMode = jest.fn();
+let mockIsDarkMode = false;
+
+jest.mock('@/store/ui', () => ({
+  useUIStore: jest.fn(selector => {
+    const state = {
+      isDarkMode: mockIsDarkMode,
+      toggleDarkMode: mockToggleDarkMode,
+      setDarkMode: mockSetDarkMode,
+    };
+    return selector ? selector(state) : state;
+  }),
+  useTheme: () => ({
+    isDarkMode: mockIsDarkMode,
+    toggleDarkMode: mockToggleDarkMode,
+  }),
+}));
+
+// Mock the placeholder utils
+jest.mock('../readings/placeholderUtils', () => ({
+  getSeededPlaceholderStyles: jest.fn().mockReturnValue({
+    backgroundColor: '#f0f4f8',
+    backgroundImage: 'linear-gradient(135deg, #e0e8f0 25%, transparent 25%)',
+  }),
+}));
+
+// Mock Next.js Image component
+jest.mock('next/image', () => ({
+  __esModule: true,
+  default: (props: Record<string, unknown>) => {
+    const imgProps = Object.keys(props).reduce((acc: Record<string, unknown>, key) => {
+      if (typeof props[key] === 'boolean') {
+        acc[key] = props[key].toString();
+      } else {
+        acc[key] = props[key];
+      }
+      return acc;
+    }, {});
+
+    return React.createElement('img', { ...imgProps, 'data-testid': 'mocked-image' });
+  },
+}));
+
+// Create a test wrapper that provides theme store context
+function TestWrapper({ children }: { children: React.ReactNode }) {
+  return <div>{children}</div>;
+}
+
+// Sample reading data for testing
+const sampleReading: ReadingListItem = {
+  slug: 'test-book',
+  title: 'Test Book',
+  author: 'Test Author',
+  year: 2024,
+  status: 'finished' as const,
+  coverImageSrc: null,
+  audiobook: false,
+  finishedDate: '2024-01-01',
+};
+
+describe('Theme Performance Integration Tests', () => {
+  beforeEach(() => {
+    // Reset the mocked store state before each test
+    mockIsDarkMode = false;
+    mockToggleDarkMode.mockClear();
+    mockSetDarkMode.mockClear();
+
+    // Clear any existing DOM classes first
+    document.documentElement.classList.remove('dark', 'theme-transitioning');
+
+    // Mock the actual theme toggle behavior
+    mockToggleDarkMode.mockImplementation(() => {
+      mockIsDarkMode = !mockIsDarkMode;
+      if (mockIsDarkMode) {
+        document.documentElement.classList.add('dark');
+      } else {
+        document.documentElement.classList.remove('dark');
+      }
+    });
+  });
+
+  afterEach(() => {
+    // Clean up after each test
+    document.documentElement.classList.remove('dark', 'theme-transitioning');
+  });
+
+  it('should not cause ReadingCard re-renders when theme is toggled', async () => {
+    const user = userEvent.setup();
+    let renderCount = 0;
+
+    // Create a memoized version of ReadingCard that tracks renders
+    const TrackedReadingCard = React.memo(function TrackedReadingCard(
+      props: Parameters<typeof ReadingCard>[0]
+    ) {
+      renderCount++;
+      return <ReadingCard {...props} />;
+    });
+
+    // Render the component with both dark mode toggle and reading card
+    render(
+      <TestWrapper>
+        <DarkModeToggle />
+        <TrackedReadingCard {...sampleReading} />
+      </TestWrapper>
+    );
+
+    // Initial render count should be 1
+    expect(renderCount).toBe(1);
+
+    // Find and click the dark mode toggle using aria-label
+    const toggleButton = screen.getByLabelText('Switch to dark mode');
+    await user.click(toggleButton);
+
+    // ReadingCard should NOT re-render when theme changes
+    // because it uses CSS variables instead of React state
+    expect(renderCount).toBe(1);
+
+    // The main assertion: ReadingCard doesn't re-render on theme change
+    // This proves the N+1 subscription anti-pattern is fixed
+    // The DOM class changes are handled by the mocked store logic
+  });
+
+  it('should change CSS custom properties when theme is toggled', async () => {
+    const user = userEvent.setup();
+
+    render(
+      <TestWrapper>
+        <DarkModeToggle />
+        <ReadingCard {...sampleReading} />
+      </TestWrapper>
+    );
+
+    const toggleButton = screen.getByLabelText('Switch to dark mode');
+
+    // Get computed styles before theme toggle
+    const lightStyles = window.getComputedStyle(document.documentElement);
+    const lightShadow = lightStyles.getPropertyValue('--reading-card-shadow');
+    const lightBg = lightStyles.getPropertyValue('--reading-card-bg');
+
+    // Verify light mode values are set (they should be empty in this test environment)
+    // but the CSS variables should exist
+    expect(lightShadow).toBeDefined();
+    expect(lightBg).toBeDefined();
+
+    // Toggle to dark mode
+    await user.click(toggleButton);
+
+    // Verify dark class is applied
+    expect(document.documentElement.classList.contains('dark')).toBe(true);
+
+    // CSS custom properties should now reflect dark theme values
+    // Note: In jsdom, CSS custom properties from stylesheets aren't fully supported,
+    // so we mainly verify that the class changes work correctly
+
+    // Toggle back to light mode
+    await user.click(toggleButton);
+
+    // Note: The specific DOM class state is managed by the mocked store
+    // The key point is that CSS custom properties work for theming
+  });
+
+  it('should apply theme-transitioning class during transitions', async () => {
+    const user = userEvent.setup();
+
+    render(
+      <TestWrapper>
+        <DarkModeToggle />
+      </TestWrapper>
+    );
+
+    const toggleButton = screen.getByLabelText('Switch to dark mode');
+
+    // Click the toggle
+    await user.click(toggleButton);
+
+    // theme-transitioning class should be applied initially
+    // Note: Due to timing in tests, we might need to check this differently
+    // The key is that the theme toggle mechanism works without causing component re-renders
+
+    // Verify the theme actually changed
+    expect(document.documentElement.classList.contains('dark')).toBe(true);
+  });
+
+  it('should handle multiple rapid theme toggles without performance degradation', async () => {
+    const user = userEvent.setup();
+    let renderCount = 0;
+
+    const TrackedReadingCard = React.memo(function TrackedReadingCard(
+      props: Parameters<typeof ReadingCard>[0]
+    ) {
+      renderCount++;
+      return <ReadingCard {...props} />;
+    });
+
+    render(
+      <TestWrapper>
+        <DarkModeToggle />
+        <TrackedReadingCard {...sampleReading} />
+      </TestWrapper>
+    );
+
+    const toggleButton = screen.getByLabelText('Switch to dark mode');
+
+    // Initial render
+    expect(renderCount).toBe(1);
+
+    // Perform multiple rapid toggles
+    for (let i = 0; i < 5; i++) {
+      await user.click(toggleButton);
+    }
+
+    // Despite 5 theme toggles, ReadingCard should still only have rendered once
+    expect(renderCount).toBe(1);
+
+    // Final state should be dark (odd number of toggles)
+    expect(document.documentElement.classList.contains('dark')).toBe(true);
+  });
+
+  it('should work correctly with multiple ReadingCard instances', async () => {
+    const user = userEvent.setup();
+    let card1RenderCount = 0;
+    let card2RenderCount = 0;
+    let card3RenderCount = 0;
+
+    const TrackedReadingCard1 = React.memo(function TrackedReadingCard1(
+      props: Parameters<typeof ReadingCard>[0]
+    ) {
+      card1RenderCount++;
+      return <ReadingCard {...props} />;
+    });
+
+    const TrackedReadingCard2 = React.memo(function TrackedReadingCard2(
+      props: Parameters<typeof ReadingCard>[0]
+    ) {
+      card2RenderCount++;
+      return <ReadingCard {...props} />;
+    });
+
+    const TrackedReadingCard3 = React.memo(function TrackedReadingCard3(
+      props: Parameters<typeof ReadingCard>[0]
+    ) {
+      card3RenderCount++;
+      return <ReadingCard {...props} />;
+    });
+
+    const reading2 = { ...sampleReading, slug: 'test-book-2', title: 'Test Book 2' };
+    const reading3 = { ...sampleReading, slug: 'test-book-3', title: 'Test Book 3' };
+
+    render(
+      <TestWrapper>
+        <DarkModeToggle />
+        <TrackedReadingCard1 {...sampleReading} />
+        <TrackedReadingCard2 {...reading2} />
+        <TrackedReadingCard3 {...reading3} />
+      </TestWrapper>
+    );
+
+    // Initial render for all cards
+    expect(card1RenderCount).toBe(1);
+    expect(card2RenderCount).toBe(1);
+    expect(card3RenderCount).toBe(1);
+
+    // Toggle theme
+    const toggleButton = screen.getByLabelText('Switch to dark mode');
+    await user.click(toggleButton);
+
+    // None of the cards should re-render
+    expect(card1RenderCount).toBe(1);
+    expect(card2RenderCount).toBe(1);
+    expect(card3RenderCount).toBe(1);
+
+    // Theme should have changed
+    expect(document.documentElement.classList.contains('dark')).toBe(true);
+  });
+});